Refactor each index into a class
Instead of everything being in Index and Search this splits it up into a new class for each of the indexes making it a bit easier to reason about.
