+++ src/MessageHandler/MeshDescriptorIndexHandler.php
@@ -6,30 +6,30 @@ namespace App\MessageHandler;
 
 use App\Entity\Manager\MeshDescriptorManager;
 use App\Message\MeshDescriptorIndexRequest;
-use App\Service\Index;
+use App\Service\Index\Mesh;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 class MeshDescriptorIndexHandler implements MessageHandlerInterface
 {
     /**
-     * @var Index
+     * @var Mesh
      */
-    private $index;
+    private $meshIndex;
 
     /**
      * @var MeshDescriptorManager
      */
     private $manager;
 
-    public function __construct(Index $index, MeshDescriptorManager $manager)
+    public function __construct(Mesh $index, MeshDescriptorManager $manager)
     {
-        $this->index = $index;
+        $this->meshIndex = $index;
         $this->manager = $manager;
     }
 
     public function __invoke(MeshDescriptorIndexRequest $message)
     {
         $descriptors = $this->manager->getIliosMeshDescriptorsById($message->getDescriptorIds());
-        $this->index->indexMeshDescriptors($descriptors);
+        $this->meshIndex->index($descriptors);
     }
 }
